refactor(cli): isolate provider account replay helpers
Problem: account operation replay pushed setup_manager beyond the repository file-size and complexity guard.\nFix: move account snapshot, summary, and replay helpers into dedicated files and keep config comments within lint limits.\nVerification: go test -mod=mod ./internal/cli -run ProviderAccount -count=1; go run ./tools/repolint; git diff --check.
